+++ app/com/lynxanalytics/biggraph/controllers/SQLController.scala
@@ -119,14 +119,10 @@ case class DataFrameSpec(
     context: SQLContext,
     viewRecipes: Iterable[(String, ViewRecipe)] = List())(
       implicit dataManager: DataManager, metaManager: MetaGraphManager): spark.sql.DataFrame = {
-    for ((name, table) <- tables) {
-      table.toDF(context).registerTempTable(name)
-    }
-    for ((name, recipe) <- viewRecipes) {
-      recipe.createDataFrame(user, context).registerTempTable(name)
-    }
-    log.info(s"Trying to execute query: ${sql}")
-    context.sql(sql)
+    val dfs =
+      tables.map { case (name, table) => name -> table.toDF(context) } ++
+        viewRecipes.map { case (name, recipe) => name -> recipe.createDataFrame(user, context) }
+    DataManager.sql(context, sql, dfs.toList)
   }
 }
 case class SQLQueryRequest(dfSpec: DataFrameSpec, maxRows: Int)
@@ -271,13 +267,19 @@ case class JdbcImportRequest(
     privacy: String,
     jdbcUrl: String,
     jdbcTable: String,
-    keyColumn: String,
+    keyColumn: Option[String] = None,
+    predicates: Option[List[String]] = None,
     overwrite: Boolean,
     columnsToImport: List[String]) extends GenericImportRequest {
 
   def dataFrame(user: serving.User, context: SQLContext)(
     implicit dataManager: DataManager): spark.sql.DataFrame = {
-    JDBCUtil.read(context, jdbcUrl, jdbcTable, keyColumn)
+    JDBCUtil.read(
+      context,
+      jdbcUrl,
+      jdbcTable,
+      keyColumn.getOrElse(""),
+      predicates.getOrElse(List()))
   }
 
   def notes: String = {