+++ CHANGELOG.md
@@ -6,8 +6,14 @@ Please add changes to "master", preferably ordered by their significance. (Most 
 
 ### master
 
- - A new machine learning operations category is created and added to the toolbox. 
- 
+ - New operation _"Triadic closure"_ has been created.
+
+### 1.9.1
+
+ - Fix startup script to support RHEL 6.
+ - A new machine learning operations category is created and added to the toolbox.
+ - SQL-related bug fixes.
+
 ### 1.9.0
 
  - Global SQL-box has been added.